+++ .github/workflows/integration_test.yaml
@@ -37,7 +37,7 @@ jobs:
       trivy-exit-code: 0
       juju-channel: 3.6/stable
       channel: 1.34-strict/stable
-      modules: '["test_charm", "test_nginx", "test_s3", "test_scaling", "test_matrix_auth"]'
+      modules: '["test_charm", "test_nginx", "test_s3", "test_scaling", "test_matrix_auth", "test_moderation"]'
   allure-report:
     if: always() && !cancelled()
     needs:
