+++ tests/integration/test_scaling.py
@@ -9,16 +9,35 @@ import typing
 import pytest
 import requests
 from juju.application import Application
+from juju.errors import JujuUnitError
 from juju.model import Model
 from ops.model import ActiveStatus
 from pytest_operator.plugin import OpsTest
-
-import synapse
 
 # mypy has trouble to inferred types for variables that are initialized in subclasses.
 ACTIVE_STATUS_NAME = typing.cast(str, ActiveStatus.name)  # type: ignore
 
 logger = logging.getLogger(__name__)
+
+
+async def test_synapse_scale_blocked(synapse_app: Application):
+    """
+    arrange: build and deploy the Synapse charm.
+    act: scale Synapse.
+    assert: the Synapse application is blocked since there is no Redis integration.
+    """
+    await synapse_app.scale(2)
+
+    with pytest.raises(JujuUnitError):
+        await synapse_app.model.wait_for_idle(
+            idle_period=30, timeout=120, apps=[synapse_app.name], raise_on_blocked=True
+        )
+
+    await synapse_app.scale(1)
+
+    await synapse_app.model.wait_for_idle(
+        idle_period=30, timeout=120, apps=[synapse_app.name], status="active"
+    )
 
 
 @pytest.mark.redis
@@ -105,32 +124,3 @@ async def test_synapse_scaling_down(
             f"http://{address}:8080/", headers={"Host": synapse_app.name}, timeout=5
         )
         assert response_worker.status_code == 200
-
-
-@pytest.mark.redis
-async def test_synapse_prometheus_configured(
-    model: Model,
-    synapse_app: Application,
-    redis_app: Application,
-    get_unit_ips: typing.Callable[[str], typing.Awaitable[tuple[str, ...]]],
-):
-    """
-    arrange: integrate Synapse with Redis and scale 1 unit.
-    act:  get all unit IPs and do a http request via port 9000.
-    assert: collect metrics should work for all units.
-    """
-    await model.wait_for_idle(
-        idle_period=30,
-        apps=[synapse_app.name, redis_app.name],
-        status=ACTIVE_STATUS_NAME,
-    )
-    await synapse_app.scale(2)
-    await model.wait_for_idle(
-        idle_period=30,
-        apps=[synapse_app.name, redis_app.name],
-        status=ACTIVE_STATUS_NAME,
-    )
-    for unit_ip in await get_unit_ips(synapse_app.name):
-        response = requests.get(f"http://{unit_ip}:{synapse.SYNAPSE_EXPORTER_PORT}/", timeout=5)
-        assert response.status_code == 200
-        assert "python_gc_objects_collected_total" in response.text
